Refactor `is_today` to `heading_matches_date`
Before, the function was called `is_today`. This did not really reflect
what the function is doing and was hard to test, due to issues with
mocking `datetime.date.today`.

Now, the function name matches better what the function does. It is
also easier to test.

This refactoring needs to continue. While creating the tests I am
finding that the functions are doing to many things and are not as
clearly defined as they could be.