Mock today correctly
Before, I had issues mocking the `datetime.date.today()` function
correctly. But, I found this in the official Python docs:
https://docs.python.org/3/library/unittest.mock-examples.html#partial-mocking

It explains exactly the difficulty involved in mocking the `today`
function and gives a pretty simple solution. That solution is
implemented now.

This allow to properly test the `get_today_heading` function.

But, I am still realizing, that it might make more sense to push the
date to an even higher level. The function would be more generalized and
reusable if it just takes a date as input. The decision which date to
extract would then be outside of this function.

Of course, at some point you need to define which date is supposed to be
extracted. Then I will definitely need to mock the `today` function.