Separate function to extract day number from head
Before, the extraction of the day number was part of the function to
build the preamble. This makes it more difficult to test and handle
error cases in particular.

Now, a new function is created to extract the day number from the
heading string. This function can easily be tested.

The function `build_preamble` can now be adjusted to handle error cases
and other possible heading formats.