Refactor get_today_heading -> get_day_heading
Before, the functions name and implementation defined the date that the
heading is looked up for. This required mocking of ``date.today`` in
testing and makes it harder to reuse the function.

Now, the function is refactored to work for any given date. This makes
testing of the function simpler and allows for reuse in other
tests (e.g. for for functions that depend on extracted heading elements)