Handle missing link cases in get_first_link
Before, there was no error handling in the `get_first_link` function.
This would make the function fail or return unexpected values if the
link address could not be extracted correctly.

Now, all cases of missing elements I could come up with and for which
tests exist are handled by returning `None`.