Add linter `wemake-python-styleguide` to repo
Before, I am only using the version that I have installed on my system,
with system wide settings. This is a bit implicit and does not show
which styleguide the project is following.

Now, the linter is a dependency of the project and the configuration is
available inside the project.