Raise exception rather than returning None
Before, when no link could be extracted `None` was returned from
`get_first_link` function.

To be more consistent with how other functions handle missing elements
that where expected, this behavior has been changed to raise an
exception instead.

I am more in favor of the style of raising exceptions in general. This
makes it impossible to "forget" to handle an error. While returning
`None` could pass unnoticed until the return value is accesses in a way
that is not possible for `None`, an exceptions stops the whole program
unless it is handled explicitly.

This style is explicitly recommended in the Python docs ("EAFP" - "it is
easier to ask for forgiveness than permission").