Raise exception rather than return None
Before, `get_day_subheading_by_text` was returning `None` when no such
subheading was found. In call case of this function returning none led
to raising a `LookupError` in the caller. In another caller, returning
`None` would lead to an `AttributeError` which is handled and leads to
a manual raise of another `LookupError`.

Now, the `get_day_subheading_by_text` function raises a `LookupError`
when no heading with the given text is found. This replaces the need to
do so in the first caller. In the second caller, this propagates further
and should show a more accurate error message.