Add unique exception msg when first p too long
Before, there was only the generic exception message that "no content
could be extracted" in the `get_tweet_message` function. This would
cover cases when there really is no content in the paragraphs, but also
when the first paragraph would be so long that it already exceeds the
maximum length.

Now, a separate exception is raised in the second case (message exists
but too long). If this message is popoulated to the user, the user can
react and change the log entry accordingly to make sure that a tweet for
the day can be generated.