Add dedicated function to add tweet to history
Before, I was only relying on the logging module. This made the behavior
hard to test. The logging module is total overkill for writing a line
to a file, which is the only thing I tried to achieve.

Now, a dedicated function write a line to a file. This can easily be
tested.