Renamed was_tweet_sent_before for consistency
Before, I was using `was_tweet_sent_before` to check for the existence
of a tweet in the history file. The tweets are added to the history
by passing it to `add_tweet_to_history`. Both of these functions are
called in the `send_tweet` function.

The difference in how these functions are named makes their connection
unclear. But calling one is required for the other to find the tweet,
and thus fulfill their combined functionality in the send function --
that is to prevent duplicate tweets.

The `was_tweet_sent_before` function is now renamed to
`is_tweet_in_history`, which much closer relates to its sibling
`add_tweet_to_history`.