Add filepath argument to is_tweet_in_history
Before, the function would be fixed to check the `logtweet.LOG_FILE`
for the tweet representation. This required mocking during the tests.

Now, the function takes the filepath as an arguments that defaults to
`logtweet.LOG_FILE`. This avoids the need to mock and is more similar to
`add_tweet_to_history` function signature.