Move history entires and config passing to main
Before, the send function was handling the checking and filling of the
history. It also depended on access to the global config object.

Now, the history checks are moved to the main. Also, the main is passing
the required config info directly into the send function.

The send function only retrieves an authenticated tweepy API and sends
the tweet. Because this relies a lot on thrid party features, I find it
extremely hard to come up with tests for it. Error in the tweepy lib
are not handled and should therefore populate to the user.