Pass Bitly API key from app.main to function
Before, because of a copy and paste left over,
`content.get_tweet_content` was trying to access the config object to
extract the Bitly API key.

Now, the config access is left to the `app.main` function and the
API key is passed to `content.get_tweet_content`. It might be not pretty
because it is not immediately obvious why the function takes that value,
but it is explained in the docstring. Also, this is an optional argument
that does not need to be provided.

The prettyness is here victim to what is clean. And it would not be
clean to access the config object in a lower level function/module.
Now the function does not need to worry about where this value is coming
from. All that the function cares is that it does something different
when that value is present.

The decision to where the value is coming from is left to a higher level
(closer) to the user. This now opens the option to add multiple ways to
configure this value if this ever needs to be done (e.g. config file,
commandline, etc.). Only the `app.main` function needs to worry about
this.