Remove exceptions from docstring
Before, the OnlineLogSource initialization function docstring mentioned
exceptions that are raised by other functions.

Since this function does not ensure these exceptions are raised, this
should not be promised in the docstring. The specific exceptions are
depended on the called functions and should only be documented where
they really occur.