Simplified `make_preamble` function
Before, the builder function `make_preamble` expected a whole day header
element as input. It then called the extraction function to grab the day
number from the heading and build the preamble.

This is a bit too much. This responsibility can easily be left to the
caller. Where the information for the day is coming from is not the
concern of this function.

Now the function only takes integer days as input. The caller still
needs to be adjusted to pass that information.