Use get_log_content_from_source in main
Before, the main function issued a request with `requests.get` directly.
This was not good, because this did not include any error handling or
validation of the source string.

Now, a specific function handles the extraction of the log content based
on a given source string. Before attempting to access the source string,
it is checked that the source string is in fact a valid URL.

Furthermore, the function to request the content is now equipped with
error handling. The error handling covers cases where the source/URL can
not be reached or responds with an error status. These issues are now
turned into understandable error messages to the user.