Add default offset 0 to argparser
Before, there was no default value for offset. This required a separate
check before the date of the day of interest could be determined. This
check had been removed in the passed. Therefore, the program call
without a given offset parameter failed.

Now, the default offset is set to `0`. This means the offset argument
always contains a meaningful value and can be used with out any checks.