Make mypy strict to catch more issues
I ran into an issue where I accidentally defined a string instead of
a tuple.
I defined the variable like so `var = ("string")`. This results in
`var = "string"`. I forgot the trailing comma to define a tuple with one
value. This is correct `var = ("string",)`.

I was expecting mypy to complain when this is passed into a function
that expects a tuple. It did, but only in strict mode. Therefore,
strict mode is now on by configuration.