Test that sep len accounted for in joining
It had occurred to me that my implementation could lead to the return
exceeding the maximum length, because I had not accounted for the
length of the separator.

This test defines a max length that is just long enough for the strings
but not the separator. Thus, only the first string should be returned.

The implementation is changed to pass the test.