Fix use case tests
Because the the source content retrievers where changed to require
a valid source instance at init, the test could not create the
mock retriever objects that need to be passed to the use case.

A mock valid source was created which could be passed to the mock
retriever classes during init.

This fixed the failing tests.