Add valid source type class variable
Before, the accepted input type for the `AbstractSourceContentRetriever`
was defined directly in it's `init` method.

This required the whole `init` function to be duplicated in a subclass,
that wants a more specific valid source type as input.

Now, the accepted valid source type can be defined as a class attribute.
This means, without having to override the parent's init method, I can
change the type of valid source that is accepted.