Require ValidOnlineSource implementation
Before, only a ValidSource implementation was required for `init` of
`OnlineSourceContentRetriever`. This is not sufficient, because there
can be additional requirements for a valid online source that are not
covered yet by a ValidSource.

Now, the `init requires a more specific implementation.

Currently the specificity only requires the name. But, the
`AbstractValidOnlineSource` can now be extended to include additional
properties and abstract methods that need to be defined in concrete
implementations.